[Agent] Fix policy and dubbo log
